fixing warning in PDB driver casting long long to long; updating Silo manual with release number; updating release notes; updating release script; fixing mkinc target
git-svn-id: file:///usr/gapps/silo/svn/trunk@142 92378438-2262-0410-827d-edb41c1c42c8

+++ src/pdb_drv/pjobj.c
@@ -717,7 +717,7 @@ PJ_ReadVariable(PDBfile *file,
       char          *c_conv = NULL;
       double        *d_conv = NULL;
       long          *l_conv = NULL;
-      long          *ll_conv = NULL;
+      long long     *ll_conv = NULL;
       short         *s_conv = NULL;
       int           *i_conv = NULL;
